requirements.txt

httpx>=0.24.0
aiohttp>=3.8.0
duckdb>=0.9.0
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
textblob>=0.17.1
feedparser>=6.0.10
emoji>=2.0.0
vaderSentiment>=3.3.2
supabase>=2.0.0
